+++ cypress_shared/pages/manage/arrivalCreate.ts
@@ -1,11 +1,11 @@
-import type { Arrival, Nonarrival } from '@approved-premises/api'
+import type { Arrival } from '@approved-premises/api'
 
 import Page from '../page'
 import paths from '../../../server/paths/manage'
 
 export default class ArrivalCreatePage extends Page {
   constructor(private readonly premisesId: string, private readonly bookingId: string) {
-    super('Did the resident arrive?')
+    super('Mark the resident as arrived')
   }
 
   static visit(premisesId: string, bookingId: string): ArrivalCreatePage {
@@ -14,10 +14,6 @@ export default class ArrivalCreatePage extends Page {
   }
 
   public completeArrivalForm(arrival: Arrival, staffMemberId: string): void {
-    cy.get('input[name="arrived"][value="Yes"]').check()
-
-    cy.log('arrival', arrival)
-
     const arrivalDate = new Date(Date.parse(arrival.arrivalDate))
     const expectedDeparture = new Date(Date.parse(arrival.expectedDepartureDate))
 
@@ -37,31 +33,7 @@ export default class ArrivalCreatePage extends Page {
     cy.get('[name="arrival[submit]"]').click()
   }
 
-  public completeNonArrivalForm(nonArrival: Nonarrival): void {
-    cy.get('input[name="arrived"][value="No"]').check()
-
-    cy.log('nonArrival', nonArrival)
-
-    const date = new Date(Date.parse(nonArrival.date))
-
-    cy.get('input[name="nonArrivalDate-day"]').type(String(date.getDate()))
-    cy.get('input[name="nonArrivalDate-month"]').type(String(date.getMonth() + 1))
-    cy.get('input[name="nonArrivalDate-year"]').type(String(date.getFullYear()))
-
-    cy.get('input[type="radio"]').last().check()
-
-    cy.get('[name="nonArrival[notes]"]').type(nonArrival.notes)
-
-    cy.get('[name="nonArrival[submit]"]').click()
-  }
-
   public submitArrivalFormWithoutFields(): void {
-    cy.get('input[name="arrived"][value="Yes"]').check()
     cy.get('[name="arrival[submit]"]').click()
-  }
-
-  public submitNonArrivalFormWithoutFields(): void {
-    cy.get('input[name="arrived"][value="No"]').check()
-    cy.get('[name="nonArrival[submit]"]').click()
   }
 }
